Further restructuring of files
